+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGES
 
+### 2.6.1
+
+* `PayloadException` will be caught and return HTTP Status 413 instead.
+* `Http410`, `Http413` and `Http418` added
+
 ### 2.6.0
 
 * `HttpConnect.autoClose` added to control whether to close the response at the end.